separated parsing code
